requirements.txt
amqp==5.3.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
huggingface-hub==0.26.5
humanfriendly==10.0
idna==3.10
kombu==5.4.2
moondream==0.0.6
mpmath==1.3.0
numpy==2.2.0
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
packaging==24.2
pillow==10.4.0
prompt_toolkit==3.0.48
protobuf==5.29.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sympy==1.13.3
tokenizers==0.20.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
